requirements.txt
tensorflow-gpu==2.0